+++ src/convert/metadataConverter.ts
@@ -12,7 +12,7 @@ import {
   ZipConfig,
 } from './types';
 import { SourceComponent } from '../resolve';
-import { promises } from 'fs';
+import { promises } from 'graceful-fs';
 import { dirname, join, normalize } from 'path';
 import { ensureDirectoryExists } from '../utils/fileSystemHandler';
 import {