+++ test/resolve/sourceComponent.test.ts
@@ -25,6 +25,10 @@ import { DECOMPOSED_COMPONENT } from '../mock/registry/type-constants/decomposed
 import { COMPONENT } from '../mock/registry/type-constants/matchingContentFileConstants';
 import {
   CHILD_1_NAME,
+  CHILD_1_XML,
+  VIRTUAL_DIR,
+  COMPONENT_1_XML,
+  COMPONENT_1_XML_PATH,
   CHILD_2_NAME,
   COMPONENT_1,
   COMPONENT_1_XML_PATH,
@@ -348,6 +352,12 @@ describe('SourceComponent', () => {
 
     it('should return correct fullName', () => {
       expect(expectedChild.fullName).to.equal(expectedChild.name);
+    });
+
+    it('should parse child xml from parent xml', () => {
+      const childXml = expectedChild.parseFromParentXml(COMPONENT_1_XML);
+      expect(childXml).to.deep.equal(CHILD_1_XML);
+      expect(COMPONENT_1.parseFromParentXml(COMPONENT_1_XML)).to.deep.equal(COMPONENT_1_XML);
     });
 
     // https://github.com/forcedotcom/salesforcedx-vscode/issues/3210