+++ test/utils/fileSystemHandler.ts
@@ -8,7 +8,7 @@ import * as fsUtil from '../../src/utils/fileSystemHandler';
 import { SinonStub, createSandbox } from 'sinon';
 import { expect } from 'chai';
 import { join } from 'path';
-import * as fs from 'fs';
+import * as fs from 'graceful-fs';
 
 const env = createSandbox();